1) переименование пакетов 2) добавлен JPA
