--- src/main/java/ru/andrey_merkulov/trader/TraderApplication.java
@@ -1,15 +1,0 @@
-package ru.andrey_merkulov.trader;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@SpringBootApplication
-@EnableAspectJAutoProxy
-public class TraderApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TraderApplication.class, args);
-	}
-
-}
