--- src/main/java/ru/andrey_merkulov/trader/TraderService.java
@@ -1,16 +1,0 @@
-package ru.andrey_merkulov.trader;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class TraderService {
-
-    public String stop() {
-        return "stopped";
-    }
-
-
-    public String start() {
-        return "started";
-    }
-}
