--- src/main/java/ru/andrey_merkulov/trader/instrument/Exception/InstrumentFoundException.java
@@ -1,7 +1,0 @@
-package ru.andrey_merkulov.trader.instrument.Exception;
-
-public class InstrumentFoundException extends RuntimeException{
-    public InstrumentFoundException(String message) {
-        super(message);
-    }
-}
