--- src/main/java/ru/andrey_merkulov/trader/instrument/Exception/InstrumentNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.andrey_merkulov.trader.instrument.Exception;
-
-public class InstrumentNotFoundException extends RuntimeException{
-    public InstrumentNotFoundException(String message) {
-        super(message);
-    }
-}
