--- src/main/java/ru/andrey_merkulov/trader/instrument/Instrument.java
@@ -1,14 +1,0 @@
-package ru.andrey_merkulov.trader.instrument;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Instrument {
-    String symbol;
-    float priceScale;
-    float minOrderQty;
-}
