--- src/main/java/ru/andrey_merkulov/trader/instrument/Serveice/InstrumentService.java
@@ -1,16 +1,0 @@
-package ru.andrey_merkulov.trader.instrument.Serveice;
-
-import org.springframework.stereotype.Service;
-import ru.andrey_merkulov.trader.instrument.Instrument;
-
-import java.util.List;
-
-@Service
-public interface InstrumentService {
-
-    List<Instrument> getAll();
-    Instrument get(String symbol);
-    void add(Instrument instrument);
-    void del(String symbol);
-    void upd(Instrument instrument);
-}
