+++ src/main/java/ru/andreymerkulov/trader/TraderApplication.java
@@ -1,0 +1,15 @@
+package ru.andreymerkulov.trader;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+
+@SpringBootApplication
+@EnableAspectJAutoProxy
+public class TraderApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TraderApplication.class, args);
+	}
+
+}
