+++ src/main/java/ru/andreymerkulov/trader/instrument/dto/InstrumentDto.java
@@ -1,0 +1,14 @@
+package ru.andreymerkulov.trader.instrument.dto;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class InstrumentDto {
+    String symbol;
+    float priceScale;
+    float minOrderQty;
+}
