+++ src/main/java/ru/andreymerkulov/trader/instrument/exception/InstrumentFoundException.java
@@ -1,0 +1,7 @@
+package ru.andreymerkulov.trader.instrument.exception;
+
+public class InstrumentFoundException extends RuntimeException{
+    public InstrumentFoundException(String message) {
+        super(message);
+    }
+}
