+++ src/main/java/ru/andreymerkulov/trader/instrument/exception/InstrumentNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.andreymerkulov.trader.instrument.exception;
+
+public class InstrumentNotFoundException extends RuntimeException{
+    public InstrumentNotFoundException(String message) {
+        super(message);
+    }
+}
