+++ src/main/java/ru/andreymerkulov/trader/instrument/repository/InstrumentDAOJPA.java
@@ -1,0 +1,7 @@
+package ru.andreymerkulov.trader.instrument.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.andreymerkulov.trader.instrument.Instrument;
+
+public interface InstrumentDAOJPA extends JpaRepository<Instrument, String> {
+}
