--- src/test/java/ru/andrey_merkulov/trader/TraderApplicationTests.java
@@ -1,14 +1,0 @@
-package ru.andrey_merkulov.trader;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TraderApplicationTests {
-
-	@Test
-	void contextLoads() {
-
-	}
-
-}
