+++ src/test/java/ru/andreymerkulov/trader/TraderApplicationTests.java
@@ -1,0 +1,14 @@
+package ru.andreymerkulov.trader;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TraderApplicationTests {
+
+	@Test
+	void contextLoads() {
+
+	}
+
+}
